refactor: Move Terminal and Teamcity renderers into Output namespace
