+++ src/Bridge/Symfony/Command/Run.php
@@ -9,8 +9,8 @@ use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
-use Testo\Bridge\Symfony\TerminalRenderer;
-use Testo\Teamcity\TeamcityRenderer;
+use Testo\Output\Teamcity\TeamcityRenderer;
+use Testo\Output\Terminal\TerminalRenderer;
 
 /**
  * Executes test suites with optional filtering and custom output formatting.
