+++ src/Output/Terminal/Renderer/DotSymbol.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Testo\Bridge\Symfony\Renderer;
+namespace Testo\Output\Terminal\Renderer;
 
 /**
  * Symbols for test status representation in dots mode.
