+++ src/Output/Terminal/Renderer/OutputFormat.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Testo\Bridge\Symfony\Renderer;
+namespace Testo\Output\Terminal\Renderer;
 
 /**
  * Output format for terminal renderer.
